requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
boto3==1.34.34
botocore==1.34.34
Werkzeug==2.3.7
python-dotenv==1.0.0